backend/requirements.txt
gunicorn
django
djangorestframework
Pillow
django-imagekit
django-cors-headers
psycopg2
google-api-python-client